+++ pytest_httpx/_httpx_internals.py
@@ -4,8 +4,6 @@ from typing import (
     Sequence,
     Tuple,
     Iterable,
-    Generator,
-    AsyncGenerator,
     Optional,
     AsyncIterator,
     Iterator,
@@ -43,35 +41,14 @@ class IteratorStream(AsyncIteratorByteStream, IteratorByteStream):
         IteratorByteStream.__init__(self, stream=Stream())
 
 
-def stream(
-    data: Any, files: Any, boundary: Optional[bytes]
+def multipart_stream(
+    data: dict, files: Any, boundary: Optional[bytes]
 ) -> Union[httpx.AsyncByteStream, httpx.SyncByteStream]:
-    if files:
-        # TODO Get rid of this internal import
-        # import is performed at runtime when needed to reduce impact of internal changes in httpx
-        from httpx._multipart import MultipartStream
-
-        return MultipartStream(data=data or {}, files=files, boundary=boundary)
-
-    if isinstance(data, str):
-        warnings.warn(
-            "data parameter as str will be removed in a future version. Use text parameter instead.",
-            DeprecationWarning,
-        )
-        data = data.encode("utf-8")
-    elif data is None:
-        data = b""
-    elif isinstance(data, bytes):
-        warnings.warn(
-            "data parameter as bytes will be removed in a future version. Use content parameter instead.",
-            DeprecationWarning,
-        )
-
-    if isinstance(data, bytes):
-        return httpx.ByteStream(data)
-
     warnings.warn(
-        "data parameter as iterator will be removed in a future version. Use stream parameter instead.",
+        "data, files and boundary parameters will be removed in a future version. Use stream parameter with an instance of httpx._multipart.MultipartStream instead.",
         DeprecationWarning,
     )
-    return IteratorStream(data)
+    # import is performed at runtime when needed to reduce impact of internal changes in httpx
+    from httpx._multipart import MultipartStream
+
+    return MultipartStream(data=data or {}, files=files, boundary=boundary)